requirements.txt
firebase-admin==6.9.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-CORS==6.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
bcrypt==4.3.0